backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.115.8",
    "pydantic>2.0",
    "sqlmodel<1.0.0,>=0.0.22",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "pyjwt<3.0.0,>=2.8.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "tenacity<10.0.0,>=9.0.0",
]
